backend/requirements.txt
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
drf-spectacular>=0.26
requests>=2.31
polyline>=2.0
whitenoise>=6.6
gunicorn>=21.2
